Split browser into controller and view
