+++ zapzap/controllers/browser/__init__.py
@@ -1,0 +1,1 @@
+from .browser_controller import BrowserController
