+++ zapzap/views/browser/__init__.py
@@ -1,0 +1,1 @@
+from .browser_view import BrowserView
